Extract older address decompaction logic into the test library
* Also add roundtrip for CompactAddr
