+++ libs/cardano-ledger-core/src/Cardano/Ledger/CompactAddress.hs
@@ -1,19 +1,25 @@
 {-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE BinaryLiterals #-}
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE DeriveFunctor #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE DerivingVia #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TupleSections #-}
 {-# LANGUAGE TypeApplications #-}
 
 module Cardano.Ledger.CompactAddress
-  ( compactAddr,
+  ( Addr (..),
+    RewardAcnt (..),
+    BootstrapAddress (..),
+    compactAddr,
     decompactAddr,
     CompactAddr (..),
-    substring,
     isPayCredScriptCompactAddr,
     isBootstrapCompactAddr,
-    -- Faster Address deserialization
     decodeAddr,
     decodeAddrShort,
     decodeAddrEither,
@@ -24,10 +30,6 @@ module Cardano.Ledger.CompactAddress
     fromCborBackwardsBothAddr,
     decodeRewardAcnt,
     fromCborRewardAcnt,
-
-    -- * Exported for benchmarking only
-    fromCborCompactAddrOld,
-    decompactAddrLazy,
   )
 where
 
@@ -36,61 +38,55 @@ import Cardano.Ledger.Address
   ( Addr (..),
     BootstrapAddress (..),
     RewardAcnt (..),
-    Word7 (..),
     byron,
-    getAddr,
-    isEnterpriseAddr,
-    notBaseAddr,
     payCredIsScript,
     serialiseAddr,
-    stakeCredIsScript,
-    toWord7,
-    word7sToWord64,
   )
-import Cardano.Ledger.BaseTypes (CertIx (..), Network (..), TxIx (..), word8ToNetwork)
+import Cardano.Ledger.BaseTypes
+  ( CertIx (CertIx),
+    Network (..),
+    SlotNo (SlotNo),
+    TxIx (TxIx),
+    byronProtVer,
+    natVersion,
+  )
 import Cardano.Ledger.Binary
   ( Decoder,
-    DecoderError (..),
     FromCBOR (..),
     ToCBOR (..),
-    byronProtVer,
-    cborError,
     decodeFull',
+    ifDecoderVersionAtLeast,
   )
 import Cardano.Ledger.Credential
-  ( Credential (KeyHashObj, ScriptHashObj),
+  ( Credential (..),
     PaymentCredential,
     Ptr (..),
     StakeReference (..),
   )
-import Cardano.Ledger.Crypto (ADDRHASH)
 import Cardano.Ledger.Crypto (Crypto)
 import Cardano.Ledger.Hashes (ScriptHash (..))
 import Cardano.Ledger.Keys (KeyHash (..))
-import Cardano.Ledger.Slot (SlotNo (..))
 import Cardano.Prelude (unsafeShortByteStringIndex)
 import Control.DeepSeq (NFData)
-import Control.Monad (ap, guard, unless, when)
-import qualified Control.Monad.Fail
+import Control.Monad (guard, unless, when)
 import Control.Monad.Trans.State (StateT, evalStateT, get, modify', state)
-import qualified Data.Binary.Get as B
-import Data.Bits (Bits, clearBit, shiftL, testBit, (.&.), (.|.))
+import Data.Bits (Bits (clearBit, shiftL, testBit, (.&.), (.|.)))
 import qualified Data.ByteString as BS
-import qualified Data.ByteString.Lazy as BSL
-import Data.ByteString.Short as SBS (fromShort, index, length, toShort)
-import Data.ByteString.Short.Internal as SBS (ShortByteString (SBS))
+import Data.ByteString.Short as SBS (ShortByteString, fromShort, index, length, toShort)
 import qualified Data.ByteString.Unsafe as BS (unsafeDrop, unsafeIndex)
-import Data.Maybe (fromMaybe)
-import qualified Data.Primitive.ByteArray as BA
 import Data.Proxy (Proxy (..))
-import Data.String (fromString)
-import Data.Text (Text, unpack)
 import Data.Word (Word16, Word32, Word64, Word8)
 import GHC.Show (intToDigit)
+import GHC.Stack (HasCallStack)
 import Numeric (showIntAtBase)
 
+------------------------------------------------------------------------------------------
+-- Compact Address -----------------------------------------------------------------------
+------------------------------------------------------------------------------------------
+
 newtype CompactAddr c = UnsafeCompactAddr ShortByteString
-  deriving (Eq, Ord, NFData)
+  deriving stock (Eq, Ord)
+  deriving newtype (NFData)
 
 instance Crypto c => Show (CompactAddr c) where
   show c = show (decompactAddr c)
@@ -99,75 +95,54 @@ compactAddr :: Addr c -> CompactAddr c
 compactAddr = UnsafeCompactAddr . SBS.toShort . serialiseAddr
 {-# INLINE compactAddr #-}
 
-decompactAddr :: forall c. Crypto c => CompactAddr c -> Addr c
+decompactAddr :: forall c. (HasCallStack, Crypto c) => CompactAddr c -> Addr c
 decompactAddr (UnsafeCompactAddr sbs) =
-  case decodeAddrShort sbs of
-    Just addr -> addr
-    Nothing -> decompactAddrOld sbs
+  case decodeAddrShortEither sbs of
+    Right addr -> addr
+    Left err ->
+      error $
+        "Impossible: Malformed CompactAddr was allowed into the system. "
+          ++ " Decoder error: "
+          ++ err
 {-# INLINE decompactAddr #-}
 
-decompactAddrOld :: Crypto c => ShortByteString -> Addr c
-decompactAddrOld short = snd . unwrap "CompactAddr" $ runGetShort getShortAddr 0 short
-  where
-    -- The reason failure is impossible here is that the only way to call this code
-    -- is using a CompactAddr, which can only be constructed using compactAddr.
-    -- compactAddr serializes an Addr, so this is guaranteed to work.
-    unwrap :: forall a. Text -> Maybe a -> a
-    unwrap name = fromMaybe (error $ unpack $ "Impossible failure when decoding " <> name)
-{-# NOINLINE decompactAddrOld #-}
-
 ------------------------------------------------------------------------------------------
--- Fast Address Serializer ---------------------------------------------------------------
+-- Address Serializer --------------------------------------------------------------------
 ------------------------------------------------------------------------------------------
 
+-- | Decoder for an `Addr`. Works in all eras
 fromCborAddr :: forall c s. Crypto c => Decoder s (Addr c)
-fromCborAddr = do
-  sbs <- fromCBOR
-  decodeAddrShort @c sbs
+fromCborAddr = fst <$> fromCborBothAddr
 {-# INLINE fromCborAddr #-}
 
+-- | Returns the actual bytes that represent an addres, while ensuring that they can
+-- be decoded in any era as an `Addr` when need be.
+fromCborCompactAddr :: forall c s. Crypto c => Decoder s (CompactAddr c)
+fromCborCompactAddr = snd <$> fromCborBothAddr
+{-# INLINE fromCborCompactAddr #-}
+
+-- | This is the decoder for an address that returns both the actual `Addr` and the bytes,
+-- that it was encoded as.
 fromCborBothAddr :: forall c s. Crypto c => Decoder s (Addr c, CompactAddr c)
 fromCborBothAddr = do
   sbs <- fromCBOR
-  addr <- decodeAddrShort @c sbs
-  pure (addr, UnsafeCompactAddr sbs)
+  case decodeAddrShortEither sbs of
+    Right addr -> pure (addr, UnsafeCompactAddr sbs)
+    -- Prior to Babbage era we did not check if binary blob representing an address was
+    -- fully consumed, so unfortunately we must preserve this behavior.
+    Left err ->
+      ifDecoderVersionAtLeast
+        (natVersion @7)
+        (fail err)
+        ((,UnsafeCompactAddr sbs) <$> evalStateT (decodeAddrStateAllowLeftoverT sbs) 0)
 {-# INLINE fromCborBothAddr #-}
-
-fromCborCompactAddr :: forall c s. Crypto c => Decoder s (CompactAddr c)
-fromCborCompactAddr = do
-  -- Ensure bytes can be decoded as Addr
-  (_addr, cAddr) <- fromCborBothAddr
-  pure cAddr
-{-# INLINE fromCborCompactAddr #-}
-
--- This is a fallback deserializer that preserves old behavior. It will almost never be
--- invoked, that is why it is not inlined.
-fromCborAddrFallback :: Crypto c => ShortByteString -> Decoder s (Addr c)
-fromCborAddrFallback sbs =
-  case B.runGetOrFail getAddr $ BSL.fromStrict $ SBS.fromShort sbs of
-    Right (_remaining, _offset, value) -> pure value
-    Left (_remaining, _offset, message) ->
-      cborError (DecoderErrorCustom "Addr" $ fromString message)
-{-# NOINLINE fromCborAddrFallback #-}
-
-fromCborCompactAddrOld :: forall s c. Crypto c => Decoder s (CompactAddr c)
-fromCborCompactAddrOld = do
-  sbs <- fromCBOR
-  UnsafeCompactAddr sbs <$ fromCborAddrFallback @c sbs
-{-# INLINE fromCborCompactAddrOld #-}
 
 fromCborBackwardsBothAddr ::
   forall c s.
   Crypto c =>
   Decoder s (Addr c, CompactAddr c)
-fromCborBackwardsBothAddr = do
-  sbs <- fromCBOR
-  addr <-
-    case decodeAddrShortEither @c sbs of
-      Right a -> pure a
-      Left _err -> fromCborAddrFallback sbs
-  pure (addr, UnsafeCompactAddr sbs)
-{-# INLINE fromCborBackwardsBothAddr #-}
+fromCborBackwardsBothAddr = fromCborBothAddr
+{-# DEPRECATED fromCborBackwardsBothAddr "Use `fromCborBothAddr` instead, they are the same" #-}
 
 class AddressBuffer b where
   bufLength :: b -> Int
@@ -202,7 +177,7 @@ instance AddressBuffer BS.ByteString where
 
 -- | Address header byte truth table:
 newtype Header = Header Word8
-  deriving (Eq, Ord, Bits, Num)
+  deriving newtype (Eq, Ord, Bits, Num)
 
 instance Show Header where
   show (Header header) = ("0b" ++) . showIntAtBase 2 intToDigit header $ ""
@@ -241,7 +216,7 @@ headerIsBaseAddress = not . (`testBit` 6)
 {-# INLINE headerIsBaseAddress #-}
 
 newtype Fail a = Fail {runFail :: Either String a}
-  deriving (Functor, Applicative, Monad)
+  deriving newtype (Functor, Applicative, Monad)
 
 instance MonadFail Fail where
   fail = Fail . Left
@@ -315,25 +290,31 @@ decodeAddrStateT ::
   (Crypto c, MonadFail m, AddressBuffer b) =>
   b ->
   StateT Int m (Addr c)
-decodeAddrStateT buf = do
+decodeAddrStateT buf = decodeAddrStateAllowLeftoverT buf <* ensureBufIsConsumed "Addr" buf
+{-# INLINE decodeAddrStateT #-}
+
+-- | Just like `decodeAddrStateT`, but does not check whether the input was consumed in full.
+decodeAddrStateAllowLeftoverT ::
+  (Crypto c, MonadFail m, AddressBuffer b) =>
+  b ->
+  StateT Int m (Addr c)
+decodeAddrStateAllowLeftoverT buf = do
   guardLength "Header" 1 buf
   let header = Header $ bufUnsafeIndex buf 0
-  addr <-
-    if isByronAddress header
-      then AddrBootstrap <$> decodeBootstrapAddress buf
-      else do
-        -- Ensure there are no unexpected bytes in the header
-        unless (header .&. headerNonShelleyBits == 0)
-          $ failDecoding
-            "Shelley Address"
-          $ "Invalid header. Unused bits are not suppose to be set: " <> show header
-        -- Advance one byte for the consumed header
-        modify' (+ 1)
-        payment <- decodePaymentCredential header buf
-        staking <- decodeStakeReference header buf
-        pure $ Addr (headerNetworkId header) payment staking
-  addr <$ ensureBufIsConsumed "Addr" buf
-{-# INLINE decodeAddrStateT #-}
+  if isByronAddress header
+    then AddrBootstrap <$> decodeBootstrapAddress buf
+    else do
+      -- Ensure there are no unexpected bytes in the header
+      unless (header .&. headerNonShelleyBits == 0)
+        $ failDecoding
+          "Shelley Address"
+        $ "Invalid header. Unused bits are not suppose to be set: " <> show header
+      -- Advance one byte for the consumed header
+      modify' (+ 1)
+      payment <- decodePaymentCredential header buf
+      staking <- decodeStakeReference header buf
+      pure $ Addr (headerNetworkId header) payment staking
+{-# INLINE decodeAddrStateAllowLeftoverT #-}
 
 -- | Checks that the current offset is exactly at the end of the buffer.
 ensureBufIsConsumed ::
@@ -507,6 +488,7 @@ decodeVariableLengthWord32 ::
 decodeVariableLengthWord32 name buf = do
   off0 <- get
   let d7 = decode7BitVarLength name buf
+      {-# INLINE d7 #-}
       d7last :: Word32 -> StateT Int m Word32
       d7last acc = do
         res <- decode7BitVarLength name buf (\_ -> failDecoding name "too many bytes.") acc
@@ -515,6 +497,7 @@ decodeVariableLengthWord32 name buf = do
         unless (bufUnsafeIndex buf off0 .&. 0b01110000 == 0) $
           failDecoding name "More than 32bits was supplied"
         pure res
+      {-# INLINE d7last #-}
   d7 (d7 (d7 (d7 d7last))) 0
 {-# INLINE decodeVariableLengthWord32 #-}
 
@@ -579,172 +562,13 @@ decodeRewardAccountT buf = do
   pure $! RewardAcnt (headerNetworkId header) account
 {-# INLINE decodeRewardAccountT #-}
 
-------------------------------------------------------------------------------------------
--- Old Address Deserializer --------------------------------------------------------------
-------------------------------------------------------------------------------------------
-
--- | This lazy deserializer is kept around purely for benchmarking, so we can
--- verify that new deserializer `decodeAddrStateT` is doing the work lazily.
-decompactAddrLazy :: forall c. Crypto c => CompactAddr c -> Addr c
-decompactAddrLazy (UnsafeCompactAddr bytes) =
-  if testBit header byron
-    then AddrBootstrap $ run "byron address" 0 bytes getBootstrapAddress
-    else Addr addrNetId paycred stakecred
-  where
-    run :: forall a. Text -> Int -> ShortByteString -> GetShort a -> a
-    run name i sbs g = snd . unwrap name $ runGetShort g i sbs
-    -- The reason failure is impossible here is that the only way to call this code
-    -- is using a CompactAddr, which can only be constructed using compactAddr.
-    -- compactAddr serializes an Addr, so this is guaranteed to work.
-    unwrap :: forall a. Text -> Maybe a -> a
-    unwrap name = fromMaybe (error $ unpack $ "Impossible failure when decoding " <> name)
-    header = run "address header" 0 bytes getWord
-    addrNetId =
-      unwrap "address network id" $
-        word8ToNetwork $
-          header .&. 0x0F -- 0b00001111 is the mask for the network id
-          -- The address format is
-          -- header | pay cred | stake cred
-          -- where the header is 1 byte
-          -- the pay cred is (sizeHash (ADDRHASH crypto)) bytes
-          -- and the stake cred can vary
-    paycred = run "payment credential" 1 bytes (getPayCred header)
-    stakecred = run "staking credential" 1 bytes $ do
-      skipHash ([] @(ADDRHASH c))
-      getStakeReference header
-    skipHash :: forall proxy h. Hash.HashAlgorithm h => proxy h -> GetShort ()
-    skipHash p = skip . fromIntegral $ Hash.sizeHash p
-    skip :: Int -> GetShort ()
-    skip n = GetShort $ \i sbs ->
-      let offsetStop = i + n
-       in if offsetStop <= SBS.length sbs
-            then Just (offsetStop, ())
-            else Nothing
-
 instance Crypto c => ToCBOR (CompactAddr c) where
   toCBOR (UnsafeCompactAddr bytes) = toCBOR bytes
+  {-# INLINE toCBOR #-}
 
 instance Crypto c => FromCBOR (CompactAddr c) where
-  fromCBOR = do
-    (_addr, cAddr) <- fromCborBackwardsBothAddr
-    pure cAddr
+  fromCBOR = fromCborCompactAddr
   {-# INLINE fromCBOR #-}
-
-newtype GetShort a = GetShort {runGetShort :: Int -> ShortByteString -> Maybe (Int, a)}
-  deriving (Functor)
-
-instance Applicative GetShort where
-  pure a = GetShort $ \i _sbs -> Just (i, a)
-  (<*>) = ap
-
-instance Monad GetShort where
-  (GetShort g) >>= f = GetShort $ \i sbs ->
-    case g i sbs of
-      Nothing -> Nothing
-      Just (i', x) -> runGetShort (f x) i' sbs
-
-instance Control.Monad.Fail.MonadFail GetShort where
-  fail _ = GetShort $ \_ _ -> Nothing
-
-getShortAddr :: Crypto c => GetShort (Addr c)
-getShortAddr = do
-  header <- peekWord8
-  if testBit header byron
-    then AddrBootstrap <$> getBootstrapAddress
-    else do
-      _ <- getWord -- read past the header byte
-      let addrNetId = header .&. 0x0F -- 0b00001111 is the mask for the network id
-      case word8ToNetwork addrNetId of
-        Just n -> do
-          c <- getPayCred header
-          h <- getStakeReference header
-          pure (Addr n c h)
-        Nothing ->
-          fail $
-            concat
-              ["Address with unknown network Id. (", show addrNetId, ")"]
-
-getBootstrapAddress :: GetShort (BootstrapAddress c)
-getBootstrapAddress = do
-  bs <- getRemainingAsByteString
-  case decodeFull' byronProtVer bs of
-    Left e -> fail $ show e
-    Right r -> pure $ BootstrapAddress r
-
-getWord :: GetShort Word8
-getWord = GetShort $ \i sbs ->
-  if i < SBS.length sbs
-    then Just (i + 1, SBS.index sbs i)
-    else Nothing
-
-peekWord8 :: GetShort Word8
-peekWord8 = GetShort peek
-  where
-    peek i sbs = if i < SBS.length sbs then Just (i, SBS.index sbs i) else Nothing
-
-getRemainingAsByteString :: GetShort BS.ByteString
-getRemainingAsByteString = GetShort $ \i sbs ->
-  let l = SBS.length sbs
-   in if i < l
-        then Just (l, SBS.fromShort $ substring sbs i l)
-        else Nothing
-
-getHash :: forall a h. Hash.HashAlgorithm h => GetShort (Hash.Hash h a)
-getHash = GetShort $ \i sbs ->
-  let hashLen = Hash.sizeHash ([] @h)
-      offsetStop = i + fromIntegral hashLen
-   in if offsetStop <= SBS.length sbs
-        then do
-          hash <- Hash.hashFromBytesShort $ substring sbs i offsetStop
-          Just (offsetStop, hash)
-        else Nothing
-
--- start is the first index copied
--- stop is the index after the last index copied
-substring :: ShortByteString -> Int -> Int -> ShortByteString
-substring (SBS ba) start stop =
-  case BA.cloneByteArray (BA.ByteArray ba) start (stop - start) of
-    BA.ByteArray ba' -> SBS ba'
-
-getWord7s :: GetShort [Word7]
-getWord7s = do
-  next <- getWord
-  -- is the high bit set?
-  if testBit next 7
-    then -- if so, grab more words
-      (:) (toWord7 next) <$> getWord7s
-    else -- otherwise, this is the last one
-      pure [Word7 next]
-
-getVariableLengthWord64 :: GetShort Word64
-getVariableLengthWord64 = word7sToWord64 <$> getWord7s
-
-getPtr :: GetShort Ptr
-getPtr =
-  Ptr
-    <$> (SlotNo <$> getVariableLengthWord64)
-    <*> (TxIx . fromIntegral <$> getVariableLengthWord64)
-    <*> (CertIx . fromIntegral <$> getVariableLengthWord64)
-
-getKeyHash :: Crypto c => GetShort (Credential kr c)
-getKeyHash = KeyHashObj . KeyHash <$> getHash
-
-getScriptHash :: Crypto c => GetShort (Credential kr c)
-getScriptHash = ScriptHashObj . ScriptHash <$> getHash
-
-getStakeReference :: Crypto c => Word8 -> GetShort (StakeReference c)
-getStakeReference header = case testBit header notBaseAddr of
-  True -> case testBit header isEnterpriseAddr of
-    True -> pure StakeRefNull
-    False -> StakeRefPtr <$> getPtr
-  False -> case testBit header stakeCredIsScript of
-    True -> StakeRefBase <$> getScriptHash
-    False -> StakeRefBase <$> getKeyHash
-
-getPayCred :: Crypto c => Word8 -> GetShort (PaymentCredential c)
-getPayCred header = case testBit header payCredIsScript of
-  True -> getScriptHash
-  False -> getKeyHash
 
 -- | Efficiently check whether compated adddress is an address with a credential
 -- that is a payment script.
